Added very basic tokenizer
It's a bit hacky and you should all feel free to butcher it up but it
works and doesn't crash with all inputs i think. Defo needs a bit of
work. Thought about turning commands into enums to make finding a
command easier than using strcomp but that's up to you guys.